Update structure: move sequence logic into the scheduler following the existing patterns. Break up methods/basic.py into a couple files to keep responsibilities separate
