+++ qiskit/compiler/sequence.py
@@ -2,7 +2,7 @@
 
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2019.
+# (C) Copyright IBM 2020.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -15,20 +15,13 @@
 """
 TODO: TO be filled.
 """
-import logging
-from collections import defaultdict
 from typing import List, Optional, Union
 
-from qiskit.circuit.barrier import Barrier
 from qiskit.circuit.quantumcircuit import QuantumCircuit
-from qiskit.exceptions import QiskitError
 from qiskit.providers import BaseBackend
 from qiskit.pulse import InstructionScheduleMap, Schedule
-from qiskit.pulse.transforms import pad
-from qiskit.scheduler import ScheduleConfig
-from qiskit.scheduler.methods.basic import translate_gates_to_pulse_defs
 
-LOG = logging.getLogger(__name__)
+from qiskit.compiler.schedule import schedule
 
 
 def sequence(scheduled_circuits: Union[QuantumCircuit, List[QuantumCircuit]],
@@ -48,59 +41,5 @@ def sequence(scheduled_circuits: Union[QuantumCircuit, List[QuantumCircuit]],
 
     Returns:
         A pulse ``Schedule`` that implements the input circuit
-
-    Raises:
-        QiskitError: If ``inst_map`` and ``meas_map`` are not passed and ``backend`` is not passed
     """
-    if inst_map is None:
-        if backend is None:
-            raise QiskitError("Must supply either a backend or InstructionScheduleMap for "
-                              "scheduling passes.")
-        inst_map = backend.defaults().instruction_schedule_map
-    if meas_map is None:
-        if backend is None:
-            raise QiskitError("Must supply either a backend or a meas_map for scheduling passes.")
-        meas_map = backend.configuration().meas_map
-
-    schedule_config = ScheduleConfig(inst_map=inst_map, meas_map=meas_map)
-    circuits = scheduled_circuits if isinstance(scheduled_circuits, list) else [scheduled_circuits]
-    schedules = [_sequence(circuit, schedule_config) for circuit in circuits]
-    return schedules[0] if len(schedules) == 1 else schedules
-
-
-def _sequence(scheduled_circuit: QuantumCircuit, schedule_config: ScheduleConfig) -> Schedule:
-    """
-    Return the pulse Schedule which implements the input circuit using an "as soon as possible"
-    (asap) scheduling policy.
-
-    Circuit instructions are first each mapped to equivalent pulse
-    Schedules according to the command definition given by the schedule_config.
-
-    Args:
-        scheduled_circuit: The scheduled quantum circuit to translate.
-        schedule_config: Backend specific parameters used for building the Schedule.
-
-    Returns:
-        A schedule corresponding to the input ``circuit``.
-    """
-    # trace start times
-    qubit_time_available = defaultdict(int)
-    start_times = []
-    for inst, qubits, _ in scheduled_circuit.data:
-        start_time = qubit_time_available[qubits[0]]
-        for q in qubits:
-            if qubit_time_available[q] != start_time:
-                raise Exception("Bug in scheduling pass.")
-
-        start_times.append(start_time)
-        for q in qubits:
-            qubit_time_available[q] += inst.duration
-
-    circ_pulse_defs = translate_gates_to_pulse_defs(scheduled_circuit, schedule_config)
-    timed_schedules = [(time, cpd.schedule) for time, cpd in zip(start_times, circ_pulse_defs)
-                       if not isinstance(cpd.schedule, Barrier)]
-    # for time, sched in timed_schedules:
-    #     print(time, sched.name, sched.duration, sched.channels)
-    sched = Schedule(*timed_schedules, name=scheduled_circuit.name)
-    assert(sched.duration == scheduled_circuit.duration)
-    return pad(sched)
+    return schedule(scheduled_circuits, backend, inst_map, meas_map, 'sequence')
