+++ qiskit/scheduler/schedule_circuit.py
@@ -20,7 +20,7 @@ from qiskit.exceptions import QiskitError
 from qiskit.pulse.schedule import Schedule
 
 from qiskit.scheduler.config import ScheduleConfig
-from qiskit.scheduler.methods.basic import as_soon_as_possible, as_late_as_possible
+from qiskit.scheduler.methods import as_soon_as_possible, as_late_as_possible, sequence
 
 
 def schedule_circuit(circuit: QuantumCircuit,
@@ -37,6 +37,8 @@ def schedule_circuit(circuit: QuantumCircuit,
           qubit resource. alias: ``'asap'``)
         * ``'as_late_as_possible'``: Schedule pulses late-- keep qubits in the ground state when
           possible. (alias: ``'alap'``)
+        * ``sequence``: Schedule trivially in an ASAP fashion, but use gate durations provided by
+                        the input schedule.
 
     Args:
         circuit: The quantum circuit to translate.
@@ -53,7 +55,8 @@ def schedule_circuit(circuit: QuantumCircuit,
         'as_soon_as_possible': as_soon_as_possible,
         'asap': as_soon_as_possible,
         'as_late_as_possible': as_late_as_possible,
-        'alap': as_late_as_possible
+        'alap': as_late_as_possible,
+        'sequence': sequence
     }
     if method is None:
         method = 'as_late_as_possible'
